refactor: extract resolveToolOverrides helper
Both policy- and tool-import paths share the same override-resolution path.
